+++ Magic/CHANGELOG.md
@@ -54,6 +54,7 @@
 
  - Mob spawners are indestructible to spells
  - The Tame action will cause tamed mobs to persist
+ - Don't allow renaming spells, skills, currency or other non-wand special items on an anvil
 
 ## Balance Changes